backend/requirements.txt
# FastAPI Backend
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# Machine Learning
scikit-learn==1.4.2
numpy==1.26.4
scipy==1.13.0

# AI/LLM Integration
google-generativeai>=0.3.0

# Database
sqlalchemy>=2.0.0

# Authentication
PyJWT>=2.8.0

# Utilities
python-dotenv>=1.0.0
python-multipart>=0.0.6

# Search optimization
rapidfuzz>=3.0.0
